requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
#python-slugify==3.0.3  # https://github.com/un33k/python-slugify
Pillow==7.1.2  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.3  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
Django==2.2.13  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
#django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-bootstrap4==0.0.7
django-debug-toolbar==2.2

# Django REST Framework
#djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
#coreapi==2.3.3  # https://github.com/core-api/python-client